refactor: rename API endpoints
