+++ packages/portal/src/server-middleware/api/events/trending.js
